requirements.txt
flask==2.0.1
pymongo==4.1.1
python-dotenv==0.19.0
face-recognition==1.3.0
opencv-python==4.5.3.56
numpy==1.21.2
Werkzeug==2.0.1
requests==2.26.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
Pillow==8.3.2
Authlib==1.0.1